requirements.txt
pandas
gTTS
pydub
numpy
librosa
noisereduce
soundfile
scipy
python-pptx
nltk
langdetect
